requirements.txt
fastapi
uvicorn
mem0ai[graph]
openai
pydantic
python-dotenv